+++ packages/ui/src/carousel/use-carousel-dot-button.ts
@@ -1,4 +1,4 @@
-import type { EmblaCarouselType } from 'embla-carousel'
+import { EmblaCarouselType } from 'embla-carousel'
 import { useCallback, useEffect, useState } from 'react'
 
 type UseDotButtonType = {
@@ -7,8 +7,9 @@ type UseDotButtonType = {
   onDotButtonClick: (index: number) => void
 }
 
-const useCarouselDotButton = (
-  emblaApi: EmblaCarouselType | undefined
+export const useDotButton = (
+  emblaApi: EmblaCarouselType | undefined,
+  onButtonClick?: (emblaApi: EmblaCarouselType) => void
 ): UseDotButtonType => {
   const [selectedIndex, setSelectedIndex] = useState(0)
   const [scrollSnaps, setScrollSnaps] = useState<number[]>([])
@@ -17,8 +18,9 @@ const useCarouselDotButton = (
     (index: number) => {
       if (!emblaApi) return
       emblaApi.scrollTo(index)
+      if (onButtonClick) onButtonClick(emblaApi)
     },
-    [emblaApi]
+    [emblaApi, onButtonClick]
   )
 
   const onInit = useCallback((emblaApi: EmblaCarouselType) => {
@@ -43,5 +45,3 @@ const useCarouselDotButton = (
     onDotButtonClick,
   }
 }
-
-export default useCarouselDotButton
